moved poll client test to the views class
