+++ polls/tests/test_views.py
@@ -1,3 +1,4 @@
+from collections import OrderedDict
 from django.test import TestCase
 from unittest.mock import patch
 from polls.models import Poll, Choice
@@ -212,3 +213,30 @@ class ResultsTest(TestCase):
             choice.save()
         response = self.client.get(f'/poll/{poll.uid}/results')
         self.assertEqual(response.context['poll'].total_votes, 10)
+
+
+class PollAPITest(TestCase):
+
+    def test_can_get_poll_data(self):
+        poll = Poll.objects.create(text='My poll text')
+        choice_0 = Choice.objects.create(text='First', poll=poll, votes=5)
+        choice_1 = Choice.objects.create(text='Second', poll=poll, votes=13)
+        response = self.client.get(f'/api/v1/poll/{poll.uid}')
+        self.assertDictEqual(response.data, {
+            'id': poll.id,
+            'text': poll.text,
+            'pub_date': poll.pub_date.astimezone().isoformat(),
+            'uid': poll.uid,
+            'choices': [
+                OrderedDict(
+                    id=choice_0.id,
+                    text=choice_0.text,
+                    votes=choice_0.votes,
+                ),
+                OrderedDict(
+                    id=choice_1.id,
+                    text=choice_1.text,
+                    votes=choice_1.votes
+                )
+            ]
+        })
